refactor(okidoc-md): refactor `buildDocumentationSourceAST` logic
BREAKING CHANGE:
change custom visitor api
