+++ packages/okidoc-md/src/buildDocumentationSource/buildDocumentationSourceAST/index.js
@@ -1,0 +1,3 @@
+import buildDocumentationSourceAST from './buildDocumentationSourceAST';
+
+export default buildDocumentationSourceAST;
